% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rparticle.R
\name{rparticle-shiny}
\alias{rparticle-shiny}
\alias{rparticleOutput}
\alias{renderRparticle}
\alias{rparticle_html}
\title{Shiny bindings for rparticle}
\usage{
rparticleOutput(outputId, width = "100\%", height = "400px")

renderRparticle(expr, env = parent.frame(), quoted = FALSE)

rparticle_html(id, style, class, ...)
}
\arguments{
\item{outputId}{output variable to read from}

\item{expr}{An expression that generates a rparticle}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using rparticle within Shiny
applications and interactive Rmd documents.
}
